Simplify redundant `catch()` during CLI bundling
Remove extra "Failed to compile" message that was redundant, and inline
"Failed to compile" message into upstream errors.

Note that now _other_ errors (such as `webpack(...)` failures) will not
be prefixed with "Failed to compile". However, it should be sufficiently
obvious that they're serious failures, so we should be OK without the
extra information.

Signed-off-by: Mitchell Hentges <mhentges@spotify.com>