chore: add lightweight dev setup e2e tests
This change adds a simple cypress setup and tests to act as a simple
sanity check for the tekton plugin's development setup.  Cypress is
added along with an additional script to easily run this e2e test from
the command line and potentially CI.  All changes are currently only
applied to the tekton plugin.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>