feat(plugins): ui for dynamic-plugins-info-backend
This change adds a dynamic-plugins-info plugin which is a frontend for
the dynamic-plugins-info-backend plugin that is part of the Janus IDP
backstage-showcase instance.  The plugin implements a table view of the
plugin data and includes client-side filtering, sorting and pagination.
A dev mode is available with some test data to populate the table.
This initial implementation only covers installing the plugin
dynamically.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>