Reduce the risk of hitting rate limits from GH
It passes the ETag header to the GH API.
For unchanged resources, the GH API will return a
"304 Not Modified" HTTP status, without any response
body, which is a hint for us to use the cached resource.
To better harness memory usage, responses are cached in
a Least-Recently-Used cache with a max size and
an expiration TTL.

As a consequence, this also makes the backend API a bit faster.