fix(bulk-import): handle Location entities created from other sources like the GH discovery as Import candidates [RHIDP-4021] (#2178)
* Display the resolved GH URLs w/o placeholders

This helps troubleshoot cache hits or misses.

* Make sure not to override all request headers passed to Octokit clients

* Load potential Imports from catalog Location entities

This covers the case of repos added
from other sources like the auto-discovery plugin

* Consider only as valid Import candidates repos that are accessible from the configured GH integrations

This should solve the issue of repos added through the auto-discovery plugin
(which seems to rely anyway on the configured GH integrations).
And for the ones coming from other sources (like app-config),
we won't show up those that are not accessible from the GH integrations
(case of any public repos out there added in some app-config.yaml).
Because ultimately, the source of truth of the bulk import right now
is the repos and orgs accessible from the GH integrations.