fix(cli): add additional shared frontend packages (#2183)
This change adds a number of additional shared frontend packages to the
webpack federated module configuration to ensure that MUI v5 styles are
correctly applied when run as a dynamic frontend plugin.  This change
fixes [RHIDP-3471](https://issues.redhat.com/browse/RHIDP-3471).

This change also adds an exclusion to the sonarcloud setup, as for
maintenance purposes we want this data structure to be explicitly
defined in the code like this.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>