feat(janus-idp/cli): add `package package-dynamic-plugins` command (#2180)
* feat(cli) POC plugin image builder command

This commit adds a POC command to build a plugin registry image from a
monorepo containing one or more plugins.  The command scans the plugins
directory for suitable plugins and runs the export-dynamic command as
needed on each plugin project.  It then copies the resulting
dist-dynamic folder for each plugin into a temporary staging directory
and generates the plugin registry metadata.  Finally it uses the
temporary directory to build a from scratch image containing all of the
plugin static assets.  The "package-dynamic-plugins" command requires
the image tag to be specified on the command line.  Additionally there's
an option to preserve the temporary staging directory for
troubleshooting purposes, as well as an option to force the regeneration
of each plugin's dist-dynamic directory

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>

* add --container-tool flag to package-dynamic-plugins

This allows users to specify if podman or docker should be used

* add list of packaged packages to package-dynamic-plugins

* package-dynamic-plugins: add --use-docker insteand of --container-tool

* package-dynamic-plugins: add pre-flight check validating podman/docker presence

---------

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>
Co-authored-by: Stan Lewis <gashcrumb@gmail.com>