fix(cli): avoid packing "binary" packages
This change updates the export-dynamic-plugin command's default behavior
to widen the check used to determine if an embedded dependency needs to
be built.  If an embedded dependency is outside of the yarn workspace
and does not appear to be a source package (where "main" is a file
ending in ".ts") then it will not be rebuilt during
export-dynamic-plugin.  This caters for cases such as a backend plugin
doing a runtime import of an ESM module, the ESM module can be built and
embedded into the dynamic plugin and be available on the package path at
runtime.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>