fix(cli): redirect yarn install output (#2580)
fix(cli) redirect yarn install output

This change updates the CLI so that stdout of the `yarn install` command
is redirected to a file.  On successful completion of `yarn install` and
validation of the exported dynamic plugin the output file will be removed,
otherwise the output file will be available for troubleshooting.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>