fix(cli): add a flag to relax version checks (#2590)
This change adds an `--ignore-version-check` flag to the CLI to allow
for selectively ignoring the semver check the CLI performs when
evaluating the plugin package dependencies.  This can be used to work
around issues where a plugin has not been updated for awhile but it is
known to work because it relies on interfaces and functions that have
not changed.

Signed-off-by: Stan Lewis <gashcrumb@gmail.com>
Co-authored-by: Stan Lewis <gashcrumb@gmail.com>